No use SCK gate

+++ SMoXPROG.cpp
@@ -32,7 +32,7 @@ SetParam()
     
     switch (param) {
     case XPRG_PARAM_NVMBASE: // PDI only
-        SMoXPROG::XPRGParam.NVMBase = (uint32_t)XPRG_Body[2] << 24 | (uint32_t)XPRG_Body[3] << 16 | (uint32_t)XPRG_Body[4] << 8 | (uint32_t)XPRG_Body[5];
+        SMoXPROG::XPRGParam.NVMBase = (uint32_t)XPRG_Body[2] << 24 | (uint32_t)XPRG_Body[3] << 16 | XPRG_Body[4] << 8 | XPRG_Body[5];
         break;
     case XPRG_PARAM_EEPPAGESIZE: // PDI only
         SMoXPROG::XPRGParam.EEPageSize = XPRG_Body[2] << 8 | XPRG_Body[3];
